requirements.txt
numpy==1.13.1
pandas==0.22.0
pillow==4.3.0
tensorflow==1.7.0